Reduce linux compilation warnings
Reduce the amount of warnings displayed during linux building, by
setting them only for core/source and black_box_tests.

Warnings from ULTs will be cleared in another patch, so for now they
will be disabled.

+++ tests/black_box_tests/xello_world_icd_clang.cpp
@@ -149,7 +149,7 @@ int main(int argc, char *argv[]) {
 
     // 6. Validate
     bool outputValidationSuccessful = true;
-    for (int i = 0; i < allocSize; ++i) {
+    for (size_t i = 0; i < allocSize; ++i) {
         outputValidationSuccessful &= ((unsigned char)(initDataSrc[i] + i) == (unsigned char)readBackData[i]);
     }
     SUCCESS_OR_WARNING_BOOL(outputValidationSuccessful);
